pulsefeed/backend/pyproject.toml
[project]
name = "pulsefeed"
version = "2.0.0"
description = "PulseFeed — stripped serving API with personalized feed recommendation"
requires-python = ">=3.13"
dependencies = [
    # FastAPI
    "fastapi>=0.135.1",
    "uvicorn>=0.42.0",
    # Database + ORM
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.3.0",
    # Auth
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    # AI (Google Gemini)
    "google-genai>=1.0",
    # Utilities
    "httpx>=0.27",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
    "redis>=5.0",
    "apscheduler>=3.10.0",
]

[tool.mypy]
python_version = "3.13"
strict = false
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "psycopg2.*",
    "pgvector.*",
    "sqlalchemy.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
